requirements.txt
# Core training dependencies
torch>=2.1.0
transformers>=4.37.0
datasets>=2.16.0
accelerate>=0.26.0
peft>=0.8.0
trl>=0.7.10
bitsandbytes>=0.42.0

# Unsloth for fast fine-tuning
# Install using: pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# Or for non-Colab: pip install "unsloth[cu121-torch230] @ git+https://github.com/unslothai/unsloth.git"

# Configuration
pyyaml>=6.0.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.16.0

# Optional: TensorBoard for logging
# tensorboard>=2.15.0

# For GGUF export
# llama-cpp-python>=0.2.0
